server/c_range.c, server/init.c, server/shop.c, server/login.c, server/c_wiz.c:   Fix some format string bugs (found using the GCC format __attribute__ for some   crossfire functions). server/c_wiz.c: Add some missing [fixed] to style_info output.
git-svn-id: svn://svn.code.sf.net/p/crossfire/code/server/trunk@11550 282e977c-c81d-0410-88c4-b93c2d0d6712

+++ server/c_range.c
@@ -151,7 +151,7 @@ static void show_matching_spells(object *op, char *params) {
                 draw_ext_info_format(NDI_UNIQUE, 0, op, MSG_TYPE_COMMAND, MSG_TYPE_COMMAND_SUCCESS,
                                      "\n[fixed]%s spells %.*s <lvl> <sp>",
                                      "\n%s spells %.*s <lvl> <sp>",
-                                     tmp, 12-strlen(tmp), "              ");
+                                     tmp, (int)(12-strlen(tmp)), "              ");
             }
             draw_ext_info_format(NDI_UNIQUE, 0, op, MSG_TYPE_COMMAND, MSG_TYPE_COMMAND_SUCCESS,
                                  "[fixed]%s",
